requirements.txt
alembic==1.4.2
aniso8601==8.0.0
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dateparser==1.0.0
dnspython==1.16.0
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Jsonpify==1.5.0
flask-marshmallow==0.13.0
Flask-Migrate==2.5.3
Flask-PyMongo==2.3.0
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.3
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.6.1
marshmallow-sqlalchemy==0.23.1
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
PyJWT==1.7.1
pymongo==3.10.1
PyMySQL==0.9.3
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.4.3
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.1
regex==2020.11.13
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.17
tzlocal==2.1
urllib3==1.25.10
wcwidth==0.2.5
Werkzeug==0.16.0
xlrd==1.2.0
zipp==3.1.0
